feat: hoist tool-output signable-candidate derivation into the sdk
The fail-closed allowlists, field-name extraction, and chain self-consistency
guards lived only in the CLI. Every other consumer would have to copy them.
Move the pure contract next to normalizeTx, keep CLI import paths as a
re-export, and take the characterization tests with the module.
