+++ packages/sdk/src/tx/index.ts
@@ -1,2 +1,16 @@
 export type { NormalizeArgs, NormalizedTx } from './normalize'
 export { normalizeTx, splitMultiTx, TxNormalizeError } from './normalize'
+export type { SignableTxCandidatePayload, ToolOutputCandidate, TxReadyPayload } from './toolOutputSigning'
+export {
+  asRecord,
+  buildTxReadyFromToolOutput,
+  buildTxReadyFromYieldOutput,
+  CLI_SIGNABLE_FLAT_TOOLS,
+  CLI_SIGNABLE_PREP_TOOLS,
+  CLI_SIGNABLE_YIELD_TOOLS,
+  deriveToolOutputCandidate,
+  DIVERGENT_FIELD_TOOLS,
+  payloadLooksSignable,
+  POLYMARKET_DEPOSIT_TOOL,
+  POLYMARKET_SETUP_TRADING_TOOL,
+} from './toolOutputSigning'
