+++ packages/sdk/tests/unit/platforms/react-native/entry.test.ts
@@ -337,6 +337,10 @@ describe('RN entry exposes pure chain helpers and registry', () => {
     expect(rn.normalizeTx).toBe(tx.normalizeTx)
     expect(rn.splitMultiTx).toBe(tx.splitMultiTx)
     expect(rn.TxNormalizeError).toBe(tx.TxNormalizeError)
+    expect(rn.deriveToolOutputCandidate).toBe(tx.deriveToolOutputCandidate)
+    expect(rn.buildTxReadyFromToolOutput).toBe(tx.buildTxReadyFromToolOutput)
+    expect(rn.payloadLooksSignable).toBe(tx.payloadLooksSignable)
+    expect(rn.CLI_SIGNABLE_FLAT_TOOLS).toBe(tx.CLI_SIGNABLE_FLAT_TOOLS)
     expect(rn.decodeFromToolResult).toBe(decode.decodeFromToolResult)
     expect(rn.decodeCosmosTx).toBe(decode.decodeCosmosTx)
     expect(rn.decodeEvmTx).toBe(decode.decodeEvmTx)
