+++ packages/sdk/tests/unit/utils/publicExports.test.ts
@@ -67,6 +67,17 @@ describe('@vultisig/sdk public exports', () => {
     expect(typeof sdk.TxNormalizeError).toBe('function')
   })
 
+  it('exports the tool-output → signable-candidate contract', () => {
+    expect(typeof sdk.deriveToolOutputCandidate).toBe('function')
+    expect(typeof sdk.buildTxReadyFromToolOutput).toBe('function')
+    expect(typeof sdk.buildTxReadyFromYieldOutput).toBe('function')
+    expect(typeof sdk.payloadLooksSignable).toBe('function')
+    expect(typeof sdk.asRecord).toBe('function')
+    expect(sdk.CLI_SIGNABLE_FLAT_TOOLS).toBeInstanceOf(Set)
+    expect(sdk.CLI_SIGNABLE_PREP_TOOLS).toBeInstanceOf(Set)
+    expect(sdk.CLI_SIGNABLE_YIELD_TOOLS).toBeInstanceOf(Set)
+  })
+
   it('exports the knownContracts canonical registry + lookup helpers', () => {
     expect(typeof sdk.isKnownContract).toBe('function')
     expect(typeof sdk.isCanonicalEvmContract).toBe('function')
